Moved scratchblocks logic into a component
